requirements.txt
psycopg2
fastapi
SQLAlchemy
uvicorn
python-dotenv
pydantic-settings>=2.0.0
pydantic[email]
passlib
passlib[bcrypt]
python-multipart
PyJWT
